structural update: InputAudioDevice and JwsBotClientSocket classes moved, NoSpeechDevice object added
