+++ .gitignore
@@ -3,6 +3,7 @@
 target
 hs_*.log
 venv
+*local*
 .DS_Store
 client/custom
 release.properties
