+++ client/common/src/main/kotlin/org/promethist/client/audio/NoSpeechDevice.kt
@@ -1,0 +1,7 @@
+package org.promethist.client.audio
+
+object NoSpeechDevice : SpeechDevice {
+    override val isSpeechDetected = false
+    override val speechAngle = -1
+    override fun close() {}
+}
