+++ client/standalone/src/main/kotlin/org/promethist/client/standalone/cli/ClientCommand.kt
@@ -17,7 +17,7 @@ import org.promethist.client.BotConfig
 import org.promethist.client.BotContext
 import org.promethist.client.audio.SpeechDevice
 import org.promethist.client.audio.WavFileAudioRecorder
-import org.promethist.client.client.JwsBotClientSocket
+import org.promethist.client.common.JwsBotClientSocket
 import org.promethist.client.common.OkHttp3BotClientSocket
 import org.promethist.client.signal.SignalGroup
 import org.promethist.client.signal.SignalProvider
